ci-automation: foundation for package + sdk update
This PR adds foundational functions to ci-automation for CI to update
ebuilds which use CROS_WORKON_COMMIT, as well as for updating the SDK
container image (based on an existing SDK container).

Updating the SDK container (instead of building a new SDK from scratch)
is menat to be used for patch-level changes which do not affect OS
images built with the SDK. For example, an SDK's mantle suite (kola
etc.) can be updated this way.

To update mantle, higher-level CI would run:
1. ci-automation/util/update_cros_workon.sh \
   sdk_container/src/third_party/coreos-overlay/coreos-devel/mantle/
2. source ci-automation/sdk_container_update.sh
3. sdk_container_update

Signed-off-by: Thilo Fromm <thilo@kinvolk.io>