version.txt: Fix SDK version (to 3165.0.0)
Release automation does not yet update the SDK version on release. It
remains stuck on the last nightly version of the main branch when a new
Alpha is released. This change updates the SDK version manually; the
root cause needs to be addressed in the release script helpers.

The change also updates coreos-overlay to the commit published with
alpha-3165.0.0.

Signed-off-by: Thilo Fromm <thilo@kinvolk.io>