sys-apps/portage: refresh patches and ensure they are applied
Found this while checking why I was still seeing lots of

  !!! Section 'gentoo' in repos.conf is missing location attribute

messages while building. Turns out that after the last sync of portage we
stopped applying patches from files/. This was caused by a local variable
definition of PATCHES that was overriding the global one.

This might be a sign to drop them or we can refresh them, as they do fix bugs
that have been hit in CoreOS in the past. I opted to refresh them, and inject
them into the local variable.