sys-apps/baselayout: Bump package version to force rebuild
The SDK container includes pre-build boards which have the basic
packages installed in the versions when the SDK was built. However,
when a change was done to one of these packages, like "baselayout",
this change is ignored unless a package version bump was done.

Bump the package version to enforce the rebuild when using the SDK
container.