prod_image_util: extract gcc libs to /usr/lib64
This made no difference back when lib was a symlink to lib64, but now that they are separate,
libs belongs in /usr/lib64. This  mostly doesn't show up because ldconfig configures the ld.so cache
to include both locations, but when updating from an older release ld.so.cache is out of date.
Unfortunately ld.so.cache does not get updated until after multipathd, which causes
multipathd to dump core. This may also affect other packages that need access to
libgcc early.

See also: https://github.com/flatcar-linux/Flatcar/issues/809