github: Fix package check
Packages in the list are not necessarily packages only, which are
represented as directories (like sys-apps/systemd), but also, in case
of eclasses, plain files. The check was checking for the path to be a
directory and emitted the warning if it was not, which resulted in
eclasses being kept not updated. Just check if the path exists.