ci-automation: Add a trap for handling flaky setups
Vendor tests are usually doing some kind of a setup before invoking
kola tests. The setup may be downloading some image files, uploading
an image to the remote service or other steps. If anything in the
setup fails then the tapfile helper that runs right after the vendor
tests fails too, because there's no tapfile created. At this point the
vendor test will not be rerun, even if it could, because the failure
during the setup was transient.

This commit adds a function that will be invoked by `trap ERR` that
will try to create a tapfile if it's missing. The tapfile will have
all the possible tests that were supposed to run marked as failure, so
the tapfile helper can parse it properly and the retest cycle can
resume.