ci-automation: Add a default kola test runner function
The vendor tests the opted in for the flaky setup handling defined
some output variables that could be then used for invoking the
run_kola_tests_on_instances function, thus making the invocation to be
exactly the same for each of those vendor tests. It gives us an
opportunity to reduce the amount of the duplicated code.