profiles: Keep using bzip2 for compressing binary packages
This was a default so far, but updating portage resulted in a change
of the default binary packages compressor from bzip2 to zstd. We
eventually will want to follow this change too, but first some scripts
that assume bzip2 will need to be updated. So for now keep using bzip2
to limit the scope of updating portage.