coreos-base/coreos-init, sys-apps/baselayout: Move installing symlinks
Install some of the symlinks in the baselayout package, instead of
coreos-init. Systemd started installing its private libraries in
/usr/lib64/systemd instead of /usr/lib/systemd which led to the
situation where /usr/lib64/systemd stopped being a symlink to
/usr/lib/systemd and became a separate directory. This could break
compatibility with software that expected things to be in
/usr/lib64/systemd.

Some symlink installation stays in init, though. This is to avoid
having them installed for some oem images where they become dangling.