pr-comment-build-dispatcher.yaml: Check team membership
This change is a building block for PR builds triggered by PR commands.
It checks for membership of the commenter in the Flatcar Maintainers
team.

Note that the "issue comment" event (which is also triggered on PR
comments) is only emitted to workflows in the "main" branch. So in order
to test / trigger this workflow, a transient "PR update" event is used,
and the "commenter" user's login is hard-coded for testing.

This will be updated to using the actual commenter's login before merge.

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>