Enable elfutils flag for systemd to get coredump stacktraces in journal
When encountering https://github.com/systemd/systemd/issues/27491 no
debug info is available when the coredump file is not found (e.g.,
because it's cleaned up). Normally a stracktrace is included in the
journal but the ebuild feature is disabled.
Enable the elfutils ebuild feature for systemd to get coredump
stacktraces in the system journal.