sys-process/tini: Move to portage-stable and sync with Gentoo
It's from Gentoo commit 3d3624f4ebb26b9fc9c9ecbff7ff0c67b3f9aa4f.

We used to have it as a fork to replace cmake-based build system with
an autotools-based one. The reason was that we didn't provide cmake in
SDK, which is not true any more.