app-containers/containerd: Migrate to Type=notify in containerd.service
Race condition arises when the containerd service unit assumes services are
ready as soon as they start running, rather than when they can actually accept
socket requests. To rectify this, changing the unit to Type=notify is required,
utilizing the existing containerd support for sd_notify call after socket setup.
In addition to this, the configuration is more aligned with upstream.

https://www.freedesktop.org/software/systemd/man/systemd.service.html#Type=