sys-kernel/coreos-sources: Backport Ampere EFI fix
There was an lockup issue with newer kernels on Ampere machines, a fix was
merged and backported to v6.1, but the fix is incomplete and the follow-up fix
was not backported (because it was not marked for stable). Add the missing patch
to our kernel for testing (and later submission to v6.1.y).

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>